addresses one bug that failed to create ifcfg-eth0 for mlab boot images.
Because the variable 'interface' is decremented as each interface is processed,
by the time the primary, is_primary=True interface is reached, the device
name "eth%s" % interface, is not eth0, but, something like eth-4, or eth-12.

This patch sorts the interfaces, implicitly placing the primary (first created)
interface first.  There is a lot of room for improvement to how this script 
handles interfaces and how it chooses the primary interface.